+++ my-app/src/components/index.js
@@ -3,7 +3,8 @@ export * from './Footer/Footer';
 export * from './Form/Form';
 export * from './Header/Header';
 export * from './Input/Input';
-export * from './InputField/InputField';
+export * from './Fields/InputField/InputField';
+export * from './Fields/TextareaField/TextareaField';
 export * from './Modal/Modal';
 export * from './Notifications/SuccessNotification/SuccessNotification';
 export * from './Textarea/Textarea';
